Improved NamingStrategy support.
